requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyYAML==6.0.1
requests==2.31.0
GitPython==3.1.40
APScheduler==3.10.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
cryptography==41.0.7
